# MRG, ENH: Speed up forward computations with Numba