--- a/mne/utils/config.py
+++ b/mne/utils/config.py
@@ -117,6 +117,7 @@ known_config_types = (
     'MNE_SKIP_TESTING_DATASET_TESTS',
     'MNE_STIM_CHANNEL',
     'MNE_USE_CUDA',
+    'MNE_USE_NUMBA',
     'MNE_SKIP_FS_FLASH_CALL',
     'SUBJECTS_DIR',
 )