This forward code takes 24 sec on master and 17 sec on this PR:

<details>

```
import time
import mne
from mne.datasets import sample
data_path = sample.data_path()
subjects_dir = data_path + '/subjects'
subject = 'sample'
bem = subjects_dir + '/sample/bem/sample-5120-5120-5120-bem-sol.fif'
raw_fname = data_path + '/MEG/sample/sample_audvis_raw.fif'
trans = data_path + '/MEG/sample/sample_audvis_raw-trans.fif'
src = mne.setup_source_space(subject, spacing='oct6',
                             subjects_dir=subjects_dir, add_dist='patch')


def test():
    t0 = time.time()
    mne.make_forward_solution(raw_fname, trans=trans, src=src, bem=bem,
                              verbose=True)
    print(time.time() - t0)


if __name__ == '__main__':
    test()
```

</details>

This cHPI fitting code takes 3.5 sec on master and 2.6 sec on this PR:

<details>

```
import time
import mne
from mne.tests.test_chpi import _assert_quats
data_path = mne.datasets.testing.data_path(download=False)
chpi_fif_fname = data_path + '/SSS/test_move_anon_raw.fif'
pos_fname = data_path + '/SSS/test_move_anon_raw.pos'
mf_quats = mne.chpi.read_head_pos(pos_fname)
raw = mne.io.read_raw_fif(chpi_fif_fname, allow_maxshield='yes', preload=True)


def test():
    py_quats = mne.chpi._calculate_chpi_positions(raw, verbose=True)
    _assert_quats(py_quats, mf_quats, dist_tol=0.004, angle_tol=3.5)


if __name__ == '__main__':
    t0 = time.time()
    test()
    print(time.time() - t0)
```

</details>