@agramfort 2019-12-11
is the new code slower without numba? FYI I have just seen again a segfault on numba code so I am super eager (yet) to push numba in core blocks of MNE