@larsoner 2019-12-11
> is the new code slower without numba?

It shouldn't be. There is only one function where I actually changed the vectorization a bit (it became simpler actually) but it did not seem to affect speed in the real-world test I ran above.

> YI I have just seen again a segfault on numba code so I am super eager (yet) to push numba in core blocks of MNE

You do have to be careful about sizing and access as their checks are not perfect. But our code has been used a while, it's pretty good about complaining when it's wrong (our unit tests are good) so I say let's keep moving forward when we can get good gains like this. If we start getting segfaults, we can remove the decorator(s) where needed.