@agramfort 2019-12-11
if there is no time regression without numba installed I am ok. Can you
just check this? thx

>