@larsoner 2019-12-11
@agramfort I checked without Numba doing EEG-only and MEG-only, and they have roughly the same timings (if anything this PR is a tiny bit faster)